% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/read.atekst_function.R
\name{read.atekst}
\alias{read.atekst}
\title{Parse an atekst .txt file}
\usage{
read.atekst(file)
}
\arguments{
\item{file}{Path to a .txt-file downloaded from atekst. Typically called something like "\code{Utvalgte_dokumenter-100-01.01.2015.txt}".}
}
\description{
Parses an atekst .txt-file. The function returns a data frame with the headline, paper, date, time, mode (net/print), url, and text for each article.
}
\examples{
corpus <- read.atekst("Utvalgte_dokumenter-100-01.01.2015.txt")
save(corpus, file = "atekst-corpus.RData")
}
\keyword{atekst}
\keyword{parse}

