% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/read.atekst.dir_function.R
\name{read.atekst.dir}
\alias{read.atekst.dir}
\title{Parse all atekst .txt files in an directory}
\usage{
read.atekst.dir(dir, recursive = TRUE, cores = 1)
}
\arguments{
\item{dir}{Path to the directory}

\item{recursive}{Should it go through all subfolders?}

\item{cores}{How many cores (i.e. parallel processes) do you wish to use?}
}
\description{
This function allows you to parse all atekst-style .txt files in an directory. The function returns a data frame with the headline, paper, data, section (if any) and main text.
}
\examples{
read.atekst.dir()
}
\keyword{atekst}

