% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/read.atekst.dir_function.R
\name{read.atekst.dir}
\alias{read.atekst.dir}
\title{Parse all atekst .txt files in a directory}
\usage{
read.atekst.dir(dir, recursive = TRUE,
  regex = "^Utvalgte_dokumenter.*.txt$", parallel = FALSE, cores = 1)
}
\arguments{
\item{dir}{Directory containing atekst .txt files.}

\item{recursive}{If \code{TRUE}, the function also parses files within subfolders.}

\item{regex}{Regular expression (pattern) to use for selecting files to parse.}

\item{parallel}{If \code{TRUE} it will try to do it in parallel (using the packages \code{foreach}, \code{iterators}, \code{doParallel} and \code{parallel}).}

\item{cores}{The amount of cores to use (if \code{parallel} is \code{TRUE}).}
}
\description{
Parse all .txt files downloaded from atekst within a directory (including subfolders). It can use a pattern (regex) to identify files. The function returns a data frame with the headline, paper, date, time, mode (net/print), url, and text for each article. In order to speed it up it is possible to run it in parallel by setting \code{parallel} to \code{TRUE} and setting \code{cores}. When working with large corpuses it is recommended to run the function once and save the resulting data frame as a \code{.RData}-file. That way it can be loaded (using \code{load()}) into R in a fraction of the time it takes to parse the whole corpus.
}
\examples{
corpus <- read.atekst.dir("some/directory")
save(corpus, file = "atekst-corpus.RData")
}
\keyword{atekst}
\keyword{parse}

