% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/read.atekst_function.R
\name{read.atekst}
\alias{read.atekst}
\title{Parse an atekst .txt file}
\usage{
read.atekst(file, object = FALSE)
}
\arguments{
\item{file}{Path to atekst .txt-file. Can also be an object (i.e., already read with readLines() if object = TRUE.}

\item{object}{If have already read the file with readLines() and wish to feed the parser an R object instead, then object should be TRUE.}
}
\description{
This function allows you to parse an atekst .txt-file. The function returns a data frame with the headline, paper, data, section (if any) and main text.
}
\examples{
read.atekst()
}
\keyword{atekst}
\keyword{parse}

